GH-12: Support multiple servers/directories
Resolves https://github.com/spring-cloud-stream-app-starters/sftp/issues/12

Add support for multiple session factories, wrapped in a `DelegatingSessionFactory`.

Apply the `RotatingServerAdice` to the inbound channel adapters.

For `listOnly` and `taskLauncherOutput` use a custom advice since the channel adapters
are not used there.
In this case defer the rotation until after extracing the environment variables for the
task launch request.