Separate sftp-data-flow source
Fixes spring-cloud-stream-app-starters/sftp#31
Fixes spring-cloud-stream-app-starters/sftp#28

The original `sftp` source functionality is the same as it was,
and the parameter `task-launcher-output` is reverted to a boolean.
A new `sftp-dataflow` source is added as the recommended way to produce
a task launch requests.
The specialized source is provided as a convenience to decouple from
the legacy TLR functionality.
This depends on https://github.com/spring-cloud-stream-app-starters/core/pull/51.
For eventual support of this functionality with additional components ,
e.g., `ftp`, a separate source may not be necessary.

Enable `TriggerProperties` and update README

Remove `@Import`

Clean up poms

Update README

Clean up

* Fix `spring-configuration-metadata-whitelist.properties` for `source`
* Polishing poms for redundant dependencies